Issue #175 part 2 Move H related files to the ChoiceModels.H folder and namespace
