+++ DaySim/ChoiceModels/Actum/Models/OtherHomeBasedTourModeTimeModel.cs
@@ -6,6 +6,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 
 using System;
+using DaySim.ChoiceModels.H;
 using DaySim.Framework.ChoiceModels;
 using DaySim.Framework.Coefficients;
 using DaySim.Framework.Core;
