+++ DaySim/DomainModels/Default/Wrappers/TripWrapper.cs
@@ -10,6 +10,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using DaySim.ChoiceModels;
+using DaySim.ChoiceModels.H;
 using DaySim.Framework.Core;
 using DaySim.Framework.DomainModels.Models;
 using DaySim.Framework.DomainModels.Persisters;
